Comic Books Are Still Made By Men, For Men And About Men

The raw data behind the story "Comic Books Are Still Made By Men, For Men And About Men" https://fivethirtyeight.com/features/women-in-comic-books/. An analysis using this data was contributed by Jonathan Bouchet as a package vignette at https://fivethirtyeight-r.netlify.com/articles/comics_gender.html.

comic_characters

Format

Because of R package size restrictions, only a preview of the first 10 rows of this dataset is included; to obtain the entire dataset (23,272 rows) see Examples below. The preview is a data frame with 10 rows representing characters and 16 variables:

publisher

Comic publisher: DC Comics or Marvel

page_id

The unique identifier for that characters page within the wikia

name

The name of the character

urlslug

The unique url within the wikia that takes you to the character

id

The identity status of the character (Secret Identity, Public identity, [on marvel only: No Dual Identity])

align

If the character is Good, Bad or Neutral

eye

Eye color of the character

hair

Hair color of the character

sex

Sex of the character (e.g. Male, Female, etc.)

gsm

If the character is a gender or sexual minority (e.g. Homosexual characters, bisexual characters)

alive

If the character is alive or deceased

appearances

The number of appearances of the character in comic books (as of Sep. 2, 2014. Number will become increasingly out of date as time goes on.)

first_appearance

The month and year of the character's first appearance in a comic book, if available

month

The month of the character's first appearance in a comic book, if available

year

The year of the character's first appearance in a comic book, if available

date

The date of the character's first appearance in a comic book, if available

Source

DC Wikia https://dc.wikia.com/wiki/Main_Page and Marvel Wikia https://marvel.wikia.com/Main_Page. Characters were scraped on August 24, 2014. Appearance counts were scraped on September 2, 2014. The month and year of the first issue each character appeared in was pulled on October 6, 2014.

Examples

# To obtain the entire dataset, run the following code:
library(readr)
library(dplyr)
library(tidyr)
library(lubridate)

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:base’:
#> 
#>     date

library(janitor)

#> 
#> Attaching package: ‘janitor’

#> The following objects are masked from ‘package:stats’:
#> 
#>     chisq.test, fisher.test

# Get DC characters:
comic_characters_dc <-
  "https://github.com/fivethirtyeight/data/raw/master/comic-characters/dc-wikia-data.csv" %>%
  read_csv() %>%
  clean_names() %>%
  mutate(publisher = "DC")

#> Parsed with column specification:
#> cols(
#>   page_id = col_double(),
#>   name = col_character(),
#>   urlslug = col_character(),
#>   ID = col_character(),
#>   ALIGN = col_character(),
#>   EYE = col_character(),
#>   HAIR = col_character(),
#>   SEX = col_character(),
#>   GSM = col_character(),
#>   ALIVE = col_character(),
#>   APPEARANCES = col_double(),
#>   `FIRST APPEARANCE` = col_character(),
#>   YEAR = col_double()
#> )

# Get Marvel characters:
comic_characters_marvel <-
  "https://github.com/fivethirtyeight/data/raw/master/comic-characters/marvel-wikia-data.csv" %>%
  read_csv() %>%
  clean_names() %>%
  mutate(publisher = "Marvel")

#> Parsed with column specification:
#> cols(
#>   page_id = col_double(),
#>   name = col_character(),
#>   urlslug = col_character(),
#>   ID = col_character(),
#>   ALIGN = col_character(),
#>   EYE = col_character(),
#>   HAIR = col_character(),
#>   SEX = col_character(),
#>   GSM = col_character(),
#>   ALIVE = col_character(),
#>   APPEARANCES = col_double(),
#>   `FIRST APPEARANCE` = col_character(),
#>   Year = col_double()
#> )

# Merge two dataset and perform further data wrangling:
comic_characters <-
  comic_characters_dc %>%
  bind_rows(comic_characters_marvel) %>%
  separate(first_appearance, c("year2", "month"), ", ", remove = FALSE) %>%
  mutate(
    # If month was missing, set as January and day as 01:
    month = ifelse(is.na(month), "01", month),
    day = "01",
    # Note some years missing:
    date = ymd(paste(year, month, day, sep = "-")),
    align = factor(
      align,
      levels = c("Bad Characters", "Reformed Criminals", "Netural Characters", "Good Characters"),
      ordered = TRUE)
  ) %>%
  select(publisher, everything(), -c(year2, day))

#> Warning: Expected 2 pieces. Missing pieces filled with `NA` in 15705 rows [210, 261, 338, 364, 558, 584, 618, 643, 715, 812, 844, 890, 1115, 1201, 1287, 1352, 1355, 1570, 1676, 1764, ...].

#> Warning:  886 failed to parse.