Elitist, Superfluous, Or Popular? We Polled Americans on the Oxford Comma

The raw data behind the story "Elitist, Superfluous, Or Popular? We Polled Americans on the Oxford Comma" http://fivethirtyeight.com/datalab/elitist-superfluous-or-popular-we-polled-americans-on-the-oxford-comma/.

comma_survey

Format

A data frame with 1129 rows representing respondents and 13 variables:

respondent_id

Respondent ID

gender

Gender

age

Age

household_income

Household income bracket

education

Education level

location

Location (census region)

more_grammar_correct

In your opinion, which sentence is more grammatically correct?

heard_oxford_comma

Prior to reading about it above, had you heard of the serial (or Oxford) comma?

care_oxford_comma

How much, if at all, do you care about the use (or lack thereof) of the serial (or Oxford) comma in grammar?

write_following

How would you write the following sentence?

data_singular_plural

When faced with using the word "data", have you ever spent time considering if the word was a singular or plural noun?

care_data

How much, if at all, do you care about the debate over the use of the word "data" as a singular or plural noun?

care_proper_grammar

In your opinion, how important or unimportant is proper use of grammar?

Source

See https://github.com/fivethirtyeight/data/tree/master/comma-survey-data.