Our Guide To The Exuberant Nonsense of College Fight Songs

The data behind the story "Our Guide To The Exuberant Nonsense Of College Fight Songs" https://projects.fivethirtyeight.com/college-fight-song-lyrics/.

fight_songs

Format

A data frame with 65 rows representing college fight songs, and 23 variables:

school

school name

conference

school college football conference

song_name

song title

writers

song author(s)

year

year the song was written; some years are unknown

student_writer

TRUE if song was written by a student, FALSE if not

official_song

TRUE if song is an official fight song according to the university, FALSE if not

contest

TRUE if song was chosen as part of a contest, FALSE if not

bpm

beats per minute

sec_duration

duration of the song in seconds

fight

TRUE if song says 'fight', FALSE if not

num_fights

number of time song says 'fight'

victory

TRUE if song says 'victory', FALSE if not

win_won

TRUE if song says 'win' or 'won', FALSE if not

victory_win_won

TRUE if song says 'victory', 'win', or 'won'

rah

TRUE if song says 'rah', FALSE if not

nonsense

TRUE if song uses nonsense syllables, FALSE if not

colors

TRUE if song mentions school colors, FALSE if not

men

TRUE if song refers to a group of men, boys, sons, etc., FALSE if not

opponents

TRUE if song mentions opponents, FALSE if not

spelling

TRUE if song spells something out, FALSE if not

trope_count

total number of tropes in song

spotify_id

Spotify id for song

Source

Spotify https://www.spotify.com/us/