fivethirtyeight: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'

An R library that provides access to the code and data sets published by FiveThirtyEight https://github.com/fivethirtyeight/data. Note that while we received guidance from editors at 538, this package is not officially published by 538. Contribute to this package at https://github.com/rudeboybert/fivethirtyeight.

Examples

# Example usage:
library(fivethirtyeight)
head(bechdel)

#> # A tibble: 6 x 15
#>    year      imdb            title            test clean_test binary    budget
#>   <int>     <chr>            <chr>           <chr>     <fctr>  <chr>     <int>
#> 1  2013 tt1711425        21 & Over          notalk     notalk   FAIL  13000000
#> 2  2012 tt1343727         Dredd 3D     ok-disagree         ok   PASS  45000000
#> 3  2013 tt2024544 12 Years a Slave notalk-disagree     notalk   FAIL  20000000
#> 4  2013 tt1272878           2 Guns          notalk     notalk   FAIL  61000000
#> 5  2013 tt0453562               42             men        men   FAIL  40000000
#> 6  2013 tt1335975         47 Ronin             men        men   FAIL 225000000
#> # ... with 8 more variables: domgross <dbl>, intgross <dbl>, code <chr>,
#> #   budget_2013 <int>, domgross_2013 <dbl>, intgross_2013 <dbl>,
#> #   period_code <int>, decade_code <int>

# All information about any data set can be found in the help file:
?bechdel

#> Using development documentation for bechdel

# To view a list of all data sets:
data(package = "fivethirtyeight")

# To view a detailed list of all data sets:
vignette("fivethirtyeight", package = "fivethirtyeight")

#> Warning: vignette ‘fivethirtyeight’ not found

# Some data sets include vignettes with an example analysis:
vignette("bechdel", package = "fivethirtyeight")

#> Warning: vignette ‘bechdel’ not found

# To browse all vignettes:
browseVignettes(package = "fivethirtyeight")

#> No vignettes found by browseVignettes(package = "fivethirtyeight")