41 Percent Of Fliers Think You're Rude If You Recline Your Seat

The raw data behind the story "41 Percent Of Fliers Think You're Rude If You Recline Your Seat" https://fivethirtyeight.com/features/airplane-etiquette-recline-seat/.

flying

Format

A data frame with 1040 rows representing respondents and 27 variables:

respondent_id

RespondentID

gender

Gender

age

Age

height

Height

children_under_18

Do you have any children under 18?

household_income

Household income bracket

education

Education Level

location

Location (census region)

frequency

How often do you travel by plane?

recline_frequency

Do you ever recline your seat when you fly?

recline_obligation

Under normal circumstances, does a person who reclines their seat during a flight have any obligation to the person sitting behind them?

recline_rude

Is it rude to recline your seat on a plane?

recline_eliminate

Given the opportunity, would you eliminate the possibility of reclining seats on planes entirely?

switch_seats_friends

Is it rude to ask someone to switch seats with you in order to be closer to friends?

switch_seats_family

Is it rude to ask someone to switch seats with you in order to be closer to family?

wake_up_bathroom

Is it rude to wake a passenger up if you are trying to go to the bathroom?

wake_up_walk

Is it rude to wake a passenger up if you are trying to walk around?

baby

In general, is it rude to bring a baby on a plane?

unruly_child

In general, is it rude to knowingly bring unruly children on a plane?

two_arm_rests

In a row of three seats, who should get to use the two arm rests?

middle_arm_rest

In a row of two seats, who should get to use the middle arm rest?

shade

Who should have control over the window shade?

unsold_seat

Is it rude to move to an unsold seat on a plane?

talk_stranger

Generally speaking, is it rude to say more than a few words to the stranger sitting next to you on a plane?

get_up

On a 6 hour flight from NYC to LA, how many times is it acceptable to get up if you're not in an aisle seat?

electronics

Have you ever used personal electronics during take off or landing in violation of a flight attendant's direction?

smoked

Have you ever smoked a cigarette in an airplane bathroom when it was against the rules?

Source

SurveyMonkey survey