The FiveThirtyEight International Food Association's 2014 World Cup

The raw data behind the story "The FiveThirtyEight International Food Association's 2014 World Cup" https://fivethirtyeight.com/features/the-fivethirtyeight-international-food-associations-2014-world-cup/. For all the countries below, the response to the following question is presented: "Please rate how much you like the traditional cuisine of X"

5: I love this country's traditional cuisine. I think it's one of the best in the world.

4: I like this country's traditional cuisine. I think it's considerably above average.

3: I'm OK with this county's traditional cuisine. I think it's about average.

2: I dislike this country's traditional cuisine. I think it's considerably below average.

1: I hate this country's traditional cuisine. I think it's one of the worst in the world.

N/A: I'm unfamiliar with this country's traditional cuisine.

food_world_cup

Format

A data frame with 1373 rows representing respondents and 48 variables:

respondent_id

Respondent ID

knowledge

Generally speaking, how would you rate your level of knowledge of cuisines from different parts of the world?

interest

How much, if at all, are you interested in cuisines from different parts of the world?

gender

Gender

age

Age

household_income

Household income bracket

education

Education Level

location

Location (census region)

algeria

Cuisine of Algeria

argentina

Cuisine of Argentina

australia

Cuisine of Australia

belgium

Cuisine of Belgium

bosnia_and_herzegovina

Cuisine of Bosnia & Herzegovina

brazil

Cuisine of Brazil

cameroon

Cuisine of Cameroon

chile

Cuisine of Chile

china

Cuisine of China

colombia

Cuisine of Colombia

costa_rica

Cuisine of Costa Rica

croatia

Cuisine of Croatia

cuba

Cuisine of Cuba

ecuador

Cuisine of Ecuador

england

Cuisine of England

ethiopia

Cuisine of Ethiopia

france

Cuisine of France

germany

Cuisine of Germany

ghana

Cuisine of Ghana

greece

Cuisine of Greece

honduras

Cuisine of Honduras

india

Cuisine of India

iran

Cuisine of Iran

ireland

Cuisine of Ireland

italy

Cuisine of Italy

ivory_coast

Cuisine of Ivory Coast

japan

Cuisine of Japan

mexico

Cuisine of Mexico

nigeria

Cuisine of Nigeria

portugal

Cuisine of Portugal

russia

Cuisine of Russia

south_korea

Cuisine of South Korea

spain

Cuisine of Spain

switzerland

Cuisine of Switzerland

thailand

Cuisine of Thailand

the_netherlands

Cuisine of the Netherlands

turkey

Cuisine of Turkey

united_states

Cuisine of the United States

uruguay

Cuisine of Uruguay

vietnam

Cuisine of Vietnam

See also