The Save Ruined Relief Pitching. The Goose Egg Can Fix It.

The raw data behind the story "The Save Ruined Relief Pitching. The Goose Egg Can Fix It." https://fivethirtyeight.com/features/goose-egg-new-save-stat-relief-pitchers/.

goose

Format

Because of R package size restrictions, only a preview of the first 10 rows of this dataset is included; to obtain the entire dataset (30,533 rows) see Examples below. The preview is a data frame with 10 rows representing pitchers and 12 variables:

name

Pitcher name

year

Start year of season

team

Retrosheet team code

league

NL or AL

goose_eggs

Goose eggs

broken_eggs

Broken eggs

mehs

Mehs

league_average_gpct

League-average goose percentage

ppf

Pitcher park factor

replacement_gpct

Replacement-level goose percentage

gwar

Goose Wins Above Replacement

key_retro

Retrosheet unique player identifier

Source

Retrosheet https://www.retrosheet.org/

Examples

# To obtain the entire dataset, run the following code:
library(readr)
library(dplyr)
goose <-
  "https://github.com/fivethirtyeight/data/raw/master/goose/goose_rawdata.csv" %>%
  read_csv(na = c("", "NA"))

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   year = col_double(),
#>   team = col_character(),
#>   league = col_character(),
#>   goose_eggs = col_double(),
#>   broken_eggs = col_double(),
#>   mehs = col_double(),
#>   league_average_gpct = col_double(),
#>   ppf = col_double(),
#>   replacement_gpct = col_double(),
#>   gwar = col_double(),
#>   key_retro = col_character()
#> )