Mayweather Vs McGregor Tweets

The raw data behind the story "The Mayweather-McGregor Fight As Told Through Emojis" https://fivethirtyeight.com/?post_type=fte_features&p=161615.

mayweather_mcgregor_tweets

Format

Because of R package size restrictions, only a preview of the first 10 rows of this dataset is included; to obtain the entire dataset (12118 rows) see Examples below. A data frame with 10 rows representing tweets and 7 variables:

created_at

Time and date at which the tweet associated with the Mayweather vs. McGregor fight was sent.

emojis

Whether or not emojis were used in the tweet about the fight.

id

A numerical identifier for each individual tweet about the fight.

link

The link to the tweet about the fight on Twitter.

retweeted

Whether or not the tweet about the fight was retweeted.

screen_name

The screen name under which the tweet about the fight was posted.

text

The text contained in the tweet about the fight.

Source

This data contains 12,118 tweets that contain one or more emojis and match one or more of the following hashtags: #MayMac, #MayweatherMcGregor, #MayweatherVMcGregor, #MayweatherVsMcGregor, #McGregor and #Mayweather. Data was collected on August 27, 2017 between 12:05 a.m. and 1:15 a.m. EDT using the Twitter streaming API. https://github.com/fivethirtyeight/data/tree/master/mayweather-mcgregor

Examples

# To obtain the entire dataset, run the following code:
library(dplyr)
library(readr)

mayweather_mcgregor_tweets <-
  "https://raw.githubusercontent.com/fivethirtyeight/data/master/mayweather-mcgregor/tweets.csv" %>%
  read_csv() %>%
  mutate(
    emojis = as.logical(emojis),
    retweeted = as.logical(retweeted),
    id = as.character(id)
  )

#> Parsed with column specification:
#> cols(
#>   created_at = col_datetime(format = ""),
#>   emojis = col_logical(),
#>   id = col_double(),
#>   link = col_character(),
#>   retweeted = col_logical(),
#>   screen_name = col_character(),
#>   text = col_character()
#> )