The Media Really Started Paying Attention to Puerto Rico When Trump Did

The raw data behind the story "The Media Really Started Paying Attention to Puerto Rico When Trump Did" https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/: Mediacloud States Data.

mediacloud_states

Format

A data frame with 51 rows representing dates and 4 variables:

date

Date

texas

The number of sentences in online news which mention Texas on the specified date

puerto_rico

The number of sentences in online news which mention Puerto Rico

florida

The number of sentences in online news which mention Florida

Source

Mediacloud https://mediacloud.org/

See also

mediacloud_hurricanes, mediacloud_online_news, mediacloud_trump, tv_hurricanes, tv_hurricanes_by_network, tv_states, google_trends