Both Parties Think The Mueller Report Was Fair. They Just Completely Disagree On What It Says.

The raw data behind the story 'Both Parties Think The Mueller Report Was Fair. They Just Completely Disagree On What It Says.' https://fivethirtyeight.com/features/both-parties-think-the-mueller-report-was-fair-they-just-disagree-on-what-it-says/

mueller_approval_polls

Format

A dataset with 65 rows representing every job approval poll of Robert Mueller that we could find from when Mueller was appointed as special council on May 17, 2017 through May 3, 2019 and 12 variables

start

the start date of the poll

end

the end date of the poll

pollster

the name of the pollster

sample_size

the size of the poll sample

population

unknown

text

the text of the poll question

approve

the number of approval in the poll

disapprove

the number of disapproval in the poll

unsure

the number of unsure in the poll

approve_(republican)

the number of approval from Republican

approve_(democrat)

the number of approval from Democrat

url

the url of the poll

Source

Polls, Washington Post / ABC and Washington Post / Schar School Polls