The Complete History Of The NBA

This directory contains the data behind the interactive "The Complete History Of The NBA" http://fivethirtyeight.com/interactives/the-complete-history-of-every-nba-team-by-elo.

nba_all_elo

Format

A data frame with 63157 rows representing basketball games and 23 variables:

gameorder

Play order of game in NBA history

game_id

Unique ID for each game

lg_id

Which league the game was played in

year_id

Season id, named based on year in which the season ended

date_game

Game date

seasongame

Number of games home team has played in season

is_playoffs

Flag for playoff games

team_id

Three letter code for home team name, from Basketball Reference

fran_id

Home team franchise id. Multiple team_ids can fall under the same fran_id due to name changes or moves. Interactive is grouped by fran_id.

pts

Points scored by home team

elo_i

Home team elo entering the game

elo_n

Home team elo following the game

win_equiv

Equivalent number of wins in a 82-game season for a team of elo_n quality

opp_id

Team id of opponent

opp_fran

Franchise id of opponent

opp_pts

Points scored by opponent

opp_elo_i

Opponent elo entering the game

opp_elo_n

Opponent elo following the game

opp_win_equiv

Opponent equivalent number of wins in a 82-game season for a team of elo_n quality

opp_seasongame

Number of games opponent has played in season

game_result

Win or loss for team in the team_id column

forecast

Elo-based chances of winning for the team in the team_id column, based on elo ratings and game location

notes

Additional information

Source

Basketball Reference: http://www.basketball-reference.com/.