The Complete History Of The NBA 2017-18 NBA Predictions

The raw data behind the story "The Complete History Of The NBA" https://projects.fivethirtyeight.com/complete-history-of-the-nba/ and our "2017-18 NBA Predictions" https://projects.fivethirtyeight.com/2018-nba-predictions/

nba_carmelo

Format

Because of R package size restrictions, only a preview of the first 10 rows of this dataset is included; to obtain the entire dataset (1871 to 2017 games) see Examples below. A data frame with 10 rows representing games and 20 variables:

date

Date

season

Season year, 1947-2018

neutral

TRUE if the game was played on neutral territory, FALSE if not

playoff

TRUE if the game was a playoff game, FALSE if not

team1

The name of one participating team

team2

The name of the other participating team

elo1_pre

Team 1's Elo rating before the game

elo2_pre

Team 2's Elo rating before the game

elo_prob1

Team 1's probability of winning based on Elo rating

elo_prob2

Team 2's probability of winning based on Elo rating

elo1_post

Team 1's Elo rating after the game

elo2_post

Team 2's Elo rating after the game

carmelo1_pre

Team 1's CARMELO rating before the game

carmelo2_pre

Team 2's CARMELO rating before the game

carmelo1_post

Team 1's CARMELO rating after the game

carmelo2_post

Team 2's CARMELO rating after the game

carmelo_prob1

Team 1's probability winning based on CARMELO rating

carmelo_prob2

Team 2's probability of winning based on CARMELO rating

score1

Points scored by Team 1

score2

Points scored by Team 2

Source

See https://projects.fivethirtyeight.com/nba-model/nba_elo.csv

Examples

# To obtain the entire dataset, run the following code:
library(tidyverse)
library(janitor)
nba_carmelo <- read_csv("https://projects.fivethirtyeight.com/nba-model/nba_elo.csv") %>%
  clean_names() %>%
  mutate(
    team1 = as.factor(team1),
    team2 = as.factor(team2),
    playoff = ifelse(playoff == "t", TRUE, FALSE),
    playoff = ifelse(is.na(playoff), FALSE, TRUE),
    neutral = ifelse(neutral == 1, TRUE, FALSE)
  )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   date = col_date(format = ""),
#>   season = col_integer(),
#>   neutral = col_integer(),
#>   elo1_pre = col_double(),
#>   elo2_pre = col_double(),
#>   elo_prob1 = col_double(),
#>   elo_prob2 = col_double(),
#>   elo1_post = col_double(),
#>   elo2_post = col_double(),
#>   score1 = col_integer(),
#>   score2 = col_integer()
#> )

#> See spec(...) for full column specifications.

#> |===========                                                     |  16%    1 MB|===========                                                     |  17%    1 MB|===========                                                     |  18%    1 MB|============                                                    |  19%    1 MB|============                                                    |  19%    1 MB|=============                                                   |  20%    1 MB|=============                                                   |  21%    1 MB|==============                                                  |  21%    1 MB|==============                                                  |  22%    1 MB|===============                                                 |  23%    1 MB|===============                                                 |  23%    1 MB|================                                                |  24%    1 MB|================                                                |  25%    2 MB|================                                                |  26%    2 MB|=================                                               |  26%    2 MB|=================                                               |  27%    2 MB|==================                                              |  28%    2 MB|==================                                              |  28%    2 MB|===================                                             |  29%    2 MB|===================                                             |  30%    2 MB|====================                                            |  31%    2 MB|====================                                            |  31%    2 MB|=====================                                           |  32%    2 MB|=====================                                           |  33%    2 MB|======================                                          |  33%    2 MB|======================                                          |  34%    2 MB|======================                                          |  35%    2 MB|=======================                                         |  36%    2 MB|=======================                                         |  36%    2 MB|========================                                        |  37%    2 MB|========================                                        |  38%    3 MB|=========================                                       |  38%    3 MB|=========================                                       |  39%    3 MB|==========================                                      |  40%    3 MB|==========================                                      |  40%    3 MB|===========================                                     |  41%    3 MB|===========================                                     |  42%    3 MB|============================                                    |  43%    3 MB|============================                                    |  43%    3 MB|============================                                    |  44%    3 MB|=============================                                   |  45%    3 MB|=============================                                   |  45%    3 MB|==============================                                  |  46%    3 MB|==============================                                  |  47%    3 MB|===============================                                 |  48%    3 MB|===============================                                 |  48%    3 MB|================================                                |  49%    3 MB|================================                                |  50%    4 MB|=================================                               |  50%    4 MB|=================================                               |  51%    4 MB|=================================                               |  52%    4 MB|==================================                              |  52%    4 MB|==================================                              |  53%    4 MB|===================================                             |  54%    4 MB|===================================                             |  55%    4 MB|====================================                            |  55%    4 MB|====================================                            |  56%    4 MB|=====================================                           |  57%    4 MB|=====================================                           |  57%    4 MB|======================================                          |  58%    4 MB|======================================                          |  59%    4 MB|======================================                          |  59%    4 MB|=======================================                         |  60%    4 MB|=======================================                         |  61%    4 MB|========================================                        |  62%    4 MB|========================================                        |  62%    5 MB|=========================================                       |  63%    5 MB|=========================================                       |  64%    5 MB|==========================================                      |  64%    5 MB|==========================================                      |  65%    5 MB|===========================================                     |  66%    5 MB|===========================================                     |  67%    5 MB|============================================                    |  67%    5 MB|============================================                    |  68%    5 MB|============================================                    |  69%    5 MB|=============================================                   |  69%    5 MB|=============================================                   |  70%    5 MB|==============================================                  |  71%    5 MB|==============================================                  |  71%    5 MB|===============================================                 |  72%    5 MB|===============================================                 |  73%    5 MB|================================================                |  74%    5 MB|================================================                |  74%    5 MB|=================================================               |  75%    6 MB|=================================================               |  76%    6 MB|=================================================               |  76%    6 MB|==================================================              |  77%    6 MB|==================================================              |  78%    6 MB|===================================================             |  78%    6 MB|===================================================             |  79%    6 MB|====================================================            |  80%    6 MB|====================================================            |  81%    6 MB|=====================================================           |  81%    6 MB|=====================================================           |  82%    6 MB|======================================================          |  83%    6 MB|======================================================          |  83%    6 MB|======================================================          |  84%    6 MB|=======================================================         |  85%    6 MB|=======================================================         |  85%    6 MB|========================================================        |  86%    6 MB|========================================================        |  87%    6 MB|=========================================================       |  88%    7 MB|=========================================================       |  88%    7 MB|==========================================================      |  90%    7 MB|===========================================================     |  91%    7 MB|============================================================    |  92%    7 MB|=============================================================   |  94%    7 MB|==============================================================  |  95%    7 MB|=============================================================== |  97%    7 MB|================================================================|  98%    7 MB|================================================================|  99%    7 MB|=================================================================| 100%    7 MB