NBA Raptor

The raw data behind the story "Introducing RAPTOR, Our New Metric For The Modern NBA" https://fivethirtyeight.com/features/introducing-raptor-our-new-metric-for-the-modern-nba.

raptor_by_player

raptor_by_team

Format

2 dataframes about Raptor players and teams by era

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 32055 rows and 24 columns.

raptor_by_player

A dataframe with 20492 rows representing every player broken out by season and era and 22 variables:

era

Historical (since the 1976 ABA-NBA merger) or Modern (since 2014)

player_name

Player Name

player_id

Player ID as found on https://www.basketball-reference.com/

season

Season

poss

Possessions played

mp

Minutes played

raptor_box_offense

Points above average per 100 possessions added by player on offense, based only on box score estimate (Values are missing when era is historical)

raptor_box_defense

Points above average per 100 possessions added by player on defense, based only on box score estimate (Values are missing when era is historical)

raptor_box_total

Points above average per 100 possessions added by player, based only on box score estimate (Values are missing when era is historical)

raptor_onoff_offense

Points above average per 100 possessions added by player on offense, based only on plus-minus data (Values are missing when era is historical)

raptor_onoff_defense

Points above average per 100 possessions added by player on defense, based only on plus-minus data (Values are missing when era is historical)

raptor_onoff_total

Points above average per 100 possessions added by player, based only on plus-minus data (Values are missing when era is historical)

raptor_offense

Points above average per 100 possessions added by player on offense, using both box and on-off components

raptor_defense

Points above average per 100 possessions added by player on defense, using both box and on-off components

raptor_total

Points above average per 100 possessions added by player on both offense and defense, using both box and on-off components

war_total

Wins Above Replacement between regular season and playoffs

war_reg_season

Wins Above Replacement for regular season

war_playoffs

Wins Above Replacement for playoffs

predator_offense

Predictive points above average per 100 possessions added by player on offense

predator_defense

Predictive points above average per 100 possessions added by player on defense

predator_total

Predictive points above average per 100 possessions added by player on both offense and defense

pace_impact

Player impact on team possessions per 48 minutes

raptor_by_team

A dataframe with 32055 rows representing every team broken out by team, season, season_type, and era and 24 variables:

era

Historical (since the 1976 ABA-NBA merger) or Modern (since 2014)

player_name

Player name

player_id

Player ID as found on https://www.basketball-reference.com/

season

Season

season_type

Regular season (RS) or playoff (PO)

team

Basketball-Reference ID of team

poss

Possessions played

mp

Minutes played

raptor_box_offense

Points above average per 100 possessions added by player on offense, based only on box score estimate (Values are missing when era is historical)

raptor_box_defense

Points above average per 100 possessions added by player on defense, based only on box score estimate (Values are missing when era is historical)

raptor_box_total

Points above average per 100 possessions added by player, based only on box score estimate (Values are missing when era is historical)

raptor_onoff_offense

Points above average per 100 possessions added by player on offense, based only on plus-minus data (Values are missing when era is historical)

raptor_onoff_defense

Points above average per 100 possessions added by player on defense, based only on plus-minus data (Values are missing when era is historical)

raptor_onoff_total

Points above average per 100 possessions added by player, based only on plus-minus data (Values are missing when era is historical)

raptor_offense

Points above average per 100 possessions added by player on offense, using both box and on-off components

raptor_defense

Points above average per 100 possessions added by player on defense, using both box and on-off components

raptor_total

Points above average per 100 possessions added by player on both offense and defense, using both box and on-off components

war_total

Wins Above Replacement between regular season and playoffs

war_reg_season

Wins Above Replacement for regular season

war_playoffs

Wins Above Replacement for playoffs

predator_offense

Predictive points above average per 100 possessions added by player on offense

predator_defense

Predictive points above average per 100 possessions added by player on defense

predator_total

Predictive points above average per 100 possessions added by player on both offense and defense

pace_impact

Player impact on team possessions per 48 minutes