The Rams Are Dead To Me, So I Answered 3,352 Questions To Find A New NFL Team

The raw data behind the story "The Rams Are Dead To Me, So I Answered 3,352 Questions To Find A New NFL Team" https://fivethirtyeight.com/features/the-rams-are-dead-to-me-so-i-answered-3352-questions-to-find-a-new-team/.

nfl_fav_team

Format

A data frame with 32 rows representing National Football League teams and 17 variables:

team

Name of NFL team

fan_relations

Fan relations - Courtesy by players, coaches and front offices toward fans, and how well a team uses technology to reach them

ownership

Ownership - Honesty; loyalty to core players and the community

players

Players - Effort on the field, likability off it

future_wins

Future wins - Projected wins over next 5 seasons

bandwagon

Bandwagon Factor - Are the team's next 5 years likely to be better than their previous 5?

tradition

Tradition - Championships/division titles/wins in team's entire history

bang_buck

Bang for the buck - Wins per fan dollars spent

behavior

Behavior - Suspensions by players on team since 2007, with extra weight to transgressions vs. women

nyc_prox

Proximity to New York City

stlouis_prox

Proximity to St. Louis

afford

Affordability - Price of tickets, parking and concessions

small_market

Small Market - Size of market in terms of population, where smaller is better

stadium_exp

Stadium experience - Quality of venue; fan-friendliness of environment; frequency of game-day promotions

coaching

Coaching - Strength of on-field leadership

uniform

Uniform - Stylishness of uniform design, according to Uni Watch's Paul Lukas

big_market

Big Market - Size of market in terms of population, where bigger is better

Source

https://www.espn.com/sportsnation/teamrankings, https://www.allourideas.org/nflteampickingsample