The Rock Isn't Alone: Lots Of People Are Worried About 'The Big One'

The raw data behind the story "The Rock Isn't Alone: Lots Of People Are Worried About 'The Big One'" https://fivethirtyeight.com/features/the-rock-isnt-alone-lots-of-people-are-worried-about-the-big-one/.

san_andreas

Format

A data frame with 1013 rows representing respondents and 11 variables:

worry_general

In general, how worried are you about earthquakes?

worry_bigone

How worried are you about the "Big One," a massive, catastrophic earthquake?

will_occur

Do you think the "Big One" will occur in your lifetime?

experience

Have you ever experienced an earthquake?

prepared

Have you or anyone in your household taken any precautions for an earthquake (packed an earthquake survival kit, prepared an evacuation plan, etc.)?

fam_san_andreas

How familiar are you with the San Andreas Fault line?

fam_yellowstone

How familiar are you with the Yellowstone Supervolcano?

age

Age

female

Gender

hhold_income

How much total combined money did all members of your HOUSEHOLD earn last year?

region

US Region

Source

See https://github.com/fivethirtyeight/data/tree/master/san-andreas