Trump World Polls Dataset:

Trump World Polls Dataset:

trumpworld_polls

Format

A data frame with 32 rows representng years and 40 variables:

year

Year the poll was conducted

avg

The average percentage people who answered the poll question positively (support the president or have a favorable view of the U.S.)

canada

The percentage of people from Canada who answered the poll question positively

france

The percentage of people from who answered the poll question positively

germany

The percentage of people from who answered the poll question positively

greece

The percentage of people from who answered the poll question positively

hungary

The percentage of people from who answered the poll question positively

italy

The percentage of people from who answered the poll question positively

netherlands

The percentage of people from who answered the poll question positively

poland

The percentage of people from who answered the poll question positively

spain

The percentage of people from who answered the poll question positively

sweden

The percentage of people from who answered the poll question positively

uk

The percentage of people from who answered the poll question positively

russia

The percentage of people from who answered the poll question positively

australia

The percentage of people from who answered the poll question positively

india

The percentage of people from who answered the poll question positively

indonesia

The percentage of people from who answered the poll question positively

japan

The percentage of people from who answered the poll question positively

philippines

The percentage of people from who answered the poll question positively

south_korea

The percentage of people from who answered the poll question positively

vietnam

The percentage of people from who answered the poll question positively

israel

The percentage of people from who answered the poll question positively

jordan

The percentage of people from who answered the poll question positively

lebanon

The percentage of people from who answered the poll question positively

tunisia

The percentage of people from who answered the poll question positively

turkey

The percentage of people from who answered the poll question positively

ghana

The percentage of people from who answered the poll question positively

kenya

The percentage of people from who answered the poll question positively

nigeria

The percentage of people from who answered the poll question positively

senegal

The percentage of people from who answered the poll question positively

south_africa

The percentage of people from who answered the poll question positively

tanzania

The percentage of people from who answered the poll question positively

argentina

The percentage of people from who answered the poll question positively

brazil

The percentage of people from who answered the poll question positively

chile

The percentage of people from who answered the poll question positively

colombia

The percentage of people from who answered the poll question positively

mexico

The percentage of people from who answered the poll question positively

peru

The percentage of people from who answered the poll question positively

venezuela

The percentage of people from who answered the poll question positively

question

The item being polled. Specificially, whether respondents: Have a favorable view of the U.S. Trust the U.S. President when it comes to foreign affairs

Source

Pew Research Center http://www.pewresearch.org/fact-tank/2017/07/17/9-charts-on-how-the-world-sees-trump/

See also

trumpworld_issues

Examples

# To convert data frame to tidy data (long) format, run:
library(tidyverse)
trumpworld_polls_tidy <- gather(trumpworld_polls, country, percent_positive, 3:39)