The Most Common Unisex Names In America: Is Yours One Of Them?

The raw data behind the story "The Most Common Unisex Names In America: Is Yours One Of Them?" https://fivethirtyeight.com/features/there-are-922-unisex-names-in-america-is-yours-one-of-them/.

unisex_names

Format

A data frame with 919 rows representing names and 5 variables:

name

First names from the Social Security Administration

total

Total number of living Americans with the name

male_share

Percentage of people with the name who are male

female_share

Percentage of people with the name who are female

gap

Gap between male_share and female_share

Source

Social Security Administration https://www.ssa.gov/oact/babynames/limits.html. See https://github.com/fivethirtyeight/data/tree/master/unisex-names.