Joining The Avengers Is As Deadly As Jumping Off A Four-Story Building

The raw data behind the story "Joining The Avengers Is As Deadly As Jumping Off A Four-Story Building" https://fivethirtyeight.com/features/avengers-death-comics-age-of-ultron/.

avengers

Format

A data frame with 173 rows representing characters and 21 variables:

url

The URL of the comic character on the Marvel Wikia

name_alias

The full name or alias of the character

appearances

The number of comic books that character appeared in as of April 30

current

Is the member currently active on an avengers affiliated team?

gender

The recorded gender of the character

probationary_intro

Sometimes the character was given probationary status as an Avenger, this is the date that happened

full_reserve_avengers_intro

The month and year the character was introduced as a full or reserve member of the Avengers

year

The year the character was introduced as a full or reserve member of the Avengers

years_since_joining

2015 minus the year

honorary

The status of the avenger, if they were given "Honorary" Avenger status, if they are simply in the "Academy," or "Full" otherwise

death1

TRUE if the Avenger died, FALSE if not.

return1

TRUE if the Avenger returned from their first death, FALSE if they did not, blank if not applicable

death2

TRUE if the Avenger died a second time after their revival, FALSE if they did not, blank if not applicable

return2

TRUE if the Avenger returned from their second death, FALSE if they did not, blank if not applicable

death3

TRUE if the Avenger died a third time after their second revival, FALSE if they did not, blank if not applicable

return3

TRUE if the Avenger returned from their third death, FALSE if they did not, blank if not applicable

death4

TRUE if the Avenger died a fourth time after their third revival, FALSE if they did not, blank if not applicable

return4

TRUE if the Avenger returned from their fourth death, FALSE if they did not, blank if not applicable

death5

TRUE if the Avenger died a fifth time after their fourth revival, FALSE if they did not, blank if not applicable

return5

TRUE if the Avenger returned from their fifth death, FALSE if they did not, blank if not applicable

notes

Descriptions of deaths and resurrections.

Source

Deaths of Marvel comic book characters between the time they joined the Avengers and April 30, 2015, the week before Secret Wars #1.