Comic Books Are Still Made By Men, For Men And About Men

The raw data behind the story "Comic Books Are Still Made By Men, For Men And About Men" http://fivethirtyeight.com/features/women-in-comic-books/.

comic_characters

Format

A data frame with 23272 rows representing characters and 16 variables:

publisher

Comic publisher: DC Comics or Marvel

page_id

The unique identifier for that characters page within the wikia

name

The name of the character

urlslug

The unique url within the wikia that takes you to the character

id

The identity status of the character (Secret Identity, Public identity, [on marvel only: No Dual Identity])

align

If the character is Good, Bad or Neutral

eye

Eye color of the character

hair

Hair color of the character

sex

Sex of the character (e.g. Male, Female, etc.)

gsm

If the character is a gender or sexual minority (e.g. Homosexual characters, bisexual characters)

alive

If the character is alive or deceased

appearances

The number of appearances of the character in comic books (as of Sep. 2, 2014. Number will become increasingly out of date as time goes on.)

first_appearance

The month and year of the character's first appearance in a comic book, if available

month

The month of the character's first appearance in a comic book, if available

year

The year of the character's first appearance in a comic book, if available

date

The date of the character's first appearance in a comic book, if available

Source

DC Wikia http://dc.wikia.com/wiki/Main_Page and Marvel Wikia http://marvel.wikia.com/Main_Page. Characters were scraped on August 24, 2014. Appearance counts were scraped on September 2, 2014. The month and year of the first issue each character appeared in was pulled on October 6, 2014.