Dear Mona Followup: Where Do People Drink The Most Beer, Wine And Spirits?

The raw data behind the story "Dear Mona Followup: Where Do People Drink The Most Beer, Wine And Spirits?" http://fivethirtyeight.com/datalab/dear-mona-followup-where-do-people-drink-the-most-beer-wine-and-spirits/.

drinks

Format

A data frame with 193 rows representing countries and 5 variables:

country

country

beer_servings

Servings of beer in average serving sizes per person

spirit_servings

Servings of spirits in average serving sizes per person

wine_servings

Servings of wine in average serving sizes per person

total_litres_of_pure_alcohol

Total litres of pure alcohol per person

Source

World Health Organisation, Global Information System on Alcohol and Health (GISAH), 2010.

Examples

# To convert data frame to tidy data (long) format, run:
library(tidyverse)
library(stringr)
drinks_tidy <- drinks %>%
  gather(type, servings, -c(country, total_litres_of_pure_alcohol)) %>%
  mutate(
    type = str_sub(type, start=1, end=-10)
  ) %>%
  arrange(country, type)