fivethirtyeight: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'

An R package that provides access to the code and data sets published by FiveThirtyEight https://github.com/fivethirtyeight/data. Note that while we received guidance from editors at 538, this package is not officially published by 538. Contribute to this package at https://github.com/rudeboybert/fivethirtyeight.

Examples

# Example usage:
library(fivethirtyeight)
head(bechdel)

#> # A tibble: 6 x 15
#>    year imdb  title test  clean_test binary budget domgross intgross code 
#>   <int> <chr> <chr> <chr> <ord>      <chr>   <int>    <dbl>    <dbl> <chr>
#> 1  2013 tt17… 21 &… nota… notalk     FAIL   1.30e7 25682380   4.22e7 2013…
#> 2  2012 tt13… Dred… ok-d… ok         PASS   4.50e7 13414714   4.09e7 2012…
#> 3  2013 tt20… 12 Y… nota… notalk     FAIL   2.00e7 53107035   1.59e8 2013…
#> 4  2013 tt12… 2 Gu… nota… notalk     FAIL   6.10e7 75612460   1.32e8 2013…
#> 5  2013 tt04… 42    men   men        FAIL   4.00e7 95020213   9.50e7 2013…
#> 6  2013 tt13… 47 R… men   men        FAIL   2.25e8 38362475   1.46e8 2013…
#> # … with 5 more variables: budget_2013 <int>, domgross_2013 <dbl>,
#> #   intgross_2013 <dbl>, period_code <int>, decade_code <int>

# All information about any data set can be found in the help file:
?bechdel

# To view a list of all data sets:
data(package = "fivethirtyeight")

# To view a detailed list of all data sets:
vignette("fivethirtyeight", package = "fivethirtyeight")

#> Warning: vignette ‘fivethirtyeight’ not found

# Some data sets include vignettes with an example analysis 
# (these are in the fivethirtyeightdata package):
if (FALSE) vignette("bechdel", package = "fivethirtyeightdata")

# To browse all vignettes (these are in the fivethirtyeightdata package):
if (FALSE) browseVignettes(package = "fivethirtyeightdata")