fivethirtyeight: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'

An R library that provides access to the code and data sets published by FiveThirtyEight https://github.com/fivethirtyeight/data. Note that while we received guidance from editors at 538, this package is not officially published by 538. Contribute to this package at https://github.com/rudeboybert/fivethirtyeight.

Examples

# Example usage:
library(fivethirtyeight)
head(bechdel)

#> # A tibble: 6 x 15
#>    year imdb   title   test    clean_test binary budget domgross intgross code 
#>   <int> <chr>  <chr>   <chr>   <ord>      <chr>   <int>    <dbl>    <dbl> <chr>
#> 1  2013 tt171… 21 & O… notalk  notalk     FAIL   1.30e⁷ 25682380   4.22e⁷ 2013…
#> 2  2012 tt134… Dredd … ok-dis… ok         PASS   4.50e⁷ 13414714   4.09e⁷ 2012…
#> 3  2013 tt202… 12 Yea… notalk… notalk     FAIL   2.00e⁷ 53107035   1.59e⁸ 2013…
#> 4  2013 tt127… 2 Guns  notalk  notalk     FAIL   6.10e⁷ 75612460   1.32e⁸ 2013…
#> 5  2013 tt045… 42      men     men        FAIL   4.00e⁷ 95020213   9.50e⁷ 2013…
#> 6  2013 tt133… 47 Ron… men     men        FAIL   2.25e⁸ 38362475   1.46e⁸ 2013…
#> # ... with 5 more variables: budget_2013 <int>, domgross_2013 <dbl>,
#> #   intgross_2013 <dbl>, period_code <int>, decade_code <int>

# All information about any data set can be found in the help file:
?bechdel

#> Rendering development documentation for 'bechdel'

# To view a list of all data sets:
data(package = "fivethirtyeight")

# To view a detailed list of all data sets:
vignette("fivethirtyeight", package = "fivethirtyeight")

#> Warning: vignette ‘fivethirtyeight’ not found

# Some data sets include vignettes with an example analysis:
vignette("bechdel", package = "fivethirtyeight")

#> Warning: vignette ‘bechdel’ not found

# To browse all vignettes:
browseVignettes(package = "fivethirtyeight")

#> No vignettes found by browseVignettes(package = "fivethirtyeight")