The Twins’ Magical Run Will Likely Go Through The Yankees. (Because Of Course.)

The raw data behind the story "The Twins’ Magical Run Will Likely Go Through The Yankees. (Because Of Course.)" https://fivethirtyeight.com/features/the-yankees-always-have-the-twins-number-will-this-postseason-be-different/

quasi_winshares

Format

A data frame contains 98,796 rows representing Quasi-Win Shares(QWS), prorated to 162 team games, for every player from 1901 as of Sept. 27, 2019 and, and 16 variables:

name_common

Player name

age

Player's age on June 30 of season

player_id

Player's https://www.baseball-reference.com/ ID

year_id

Season

team_id

Baseball-Reference ID of team

lg_id

League of team

pct_pt

Share of total team playing time (measured by plate appearances and leverage-weighted innings) for player

war162

Wins Above Replacement per 162 team games, averaging B-R and FG WAR

quasi_ws

Quasi Win Shares. Three times total wins created per 162 games; generated by adding WAR162 to wins BELOW replacement (determined by playing time) and rounding to nearest whole number

stint_id

Order of team stints for player within season

franch_id

Lahman DB franchise ID for team

prev_franch

Previous franchise played for (NA if player debuted for team)

year_acq

First season for player with that franchise

year_left

Last season for player with that franchise

next_franch

Next franchise played for (NA if player retired with or is currently on team)

p

If played as a Pitcher

c

If played as a Catcher

X1b

If played as a First Base

X2b

If played as a Second Base

X3b

If played as a Third Base

ss

If played as a Shortstop

lf

If played as a Left Field

cf

If played as a Center Field

rf

If played as a Right Field

Source

What’s The Best Way To Build a Major League Baseball Team? https://fivethirtyeight.com/features/whats-the-best-way-to-build-a-major-league-baseball-team/